增加手机信息模块

+++ src/main/java/edu/xhu/mobilee/dao/ManufacturerDao.java
@@ -12,5 +12,5 @@ public interface ManufacturerDao {
 
     int deleteManufacturer(List<String> id);
 
-    int insertManufacturer(ManufacturerEntity manufacturerEntity);
+    long insertManufacturer(ManufacturerEntity manufacturerEntity);
 }
