+++ src/main/java/edu/xhu/mobilee/service/ManufacturerService.java
@@ -13,5 +13,5 @@ public interface ManufacturerService {
 
     int deleteManufacturer(List<String> id);
 
-    int insertManufacturer(ManufacturerEntity manufacturerEntity);
+    long insertManufacturer(ManufacturerEntity manufacturerEntity);
 }
